+++ History.markdown
@@ -1,5 +1,10 @@
 # Change Log
 
+## 3.0.3 / 2020-04-19
+
+Allow `site.metadata.twitter` data if present. Look for `page.subtitle` if
+`page.description` isn't present.
+
 ## 3.0.0 / 2020-04-14
 
 Use Bridgetown gem and rename to bridgetown-seo-tag.